#100 Added test for variables

+++ bin/airgapped/download-helm-charts.sh
@@ -9,6 +9,13 @@ if [[ -z "${SENZING_AIRGAPPED_DIR}" ]]; then
     echo "Error: SENZING_AIRGAPPED_DIR must be set"
 fi
 
+if [ ! -f "${SENZING_AIRGAPPED_DIR}/kubernetes-demo/helm-values-templates/senzing-stream-producer-rabbitmq.yaml" ]; then
+    ERRORS=$((${ERRORS} + 1))
+    echo "Error: SENZING_AIRGAPPED_DIR may not be set correctly."
+    echo "       Current value: ${SENZING_AIRGAPPED_DIR}"
+    echo "Error: Could not find ${SENZING_AIRGAPPED_DIR}/kubernetes-demo/helm-values-templates/senzing-stream-producer-rabbitmq.yaml"
+fi
+
 if [[ ${ERRORS} > 0 ]]; then
     echo "No processing done. ${ERRORS} errors found."
     exit 1
@@ -48,21 +55,24 @@ rm    ${SENZING_AIRGAPPED_DIR}/bitnami-charts.zip
 # Remove extraneous sub-directories.
 
 pushd ${SENZING_AIRGAPPED_DIR}/bitnami-charts
-if [[ $?!= 0 ]]; then
+returnCode=$?
+if [[ ${returnCode} -ne 0 ]]; then
     echo "Error: ${SENZING_AIRGAPPED_DIR}/bitnami-charts directory does not exist."
-    exit
+    exit 1
 fi
-
-
-echo "FIXME: Shouldn't be here"
-exit
-
 
 rm *
 rm .*
 rm -rf .*
 rm -rf !("bitnami")
+
 pushd ${SENZING_AIRGAPPED_DIR}/bitnami-charts/bitnami
+returnCode=$?
+if [[ ${returnCode} -ne 0 ]]; then
+    echo "Error: ${SENZING_AIRGAPPED_DIR}/bitnami-charts/bitnami directory does not exist."
+    exit 1
+fi
+
 rm -rf !("postgresql"|"rabbitmq")
 popd
 popd
@@ -103,12 +113,26 @@ rm    ${SENZING_AIRGAPPED_DIR}/senzing-charts.zip
 # Remove extraneous sub-directories.
 
 pushd ${SENZING_AIRGAPPED_DIR}/senzing-charts
+returnCode=$?
+if [[ ${returnCode} -ne 0 ]]; then
+    echo "Error: ${SENZING_AIRGAPPED_DIR}/senzing-charts directory does not exist."
+    exit 1
+fi
+
 rm *
 rm .*
 rm -rf .*
 rm -rf !("charts")
+
 pushd ${SENZING_AIRGAPPED_DIR}/senzing-charts/charts
+returnCode=$?
+if [[ ${returnCode} -ne 0 ]]; then
+    echo "Error: ${SENZING_AIRGAPPED_DIR}/senzing-charts/charts directory does not exist."
+    exit 1
+fi
+
 rm -rf !("phppgadmin"|"senzing-api-server"|"senzing-configurator"|"senzing-console"|"senzing-entity-search-web-app"|"senzing-init-container"|"senzing-installer"|"senzing-postgresql-client"|"senzing-redoer"|"senzing-stream-loader"|"senzing-stream-producer"|"swaggerapi-swagger-ui"|)
+
 popd
 popd
